requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
pyyaml>=6.0
ollama>=0.3.0
tweepy>=4.14.0
